+++ OmeSliCC/OmeroSource.py
@@ -81,11 +81,6 @@ class OmeroSource(OmeSource):
         self.source_pixel_size = [(get_default(image_object.getPixelSizeX(), 0), self.default_physical_unit),
                                   (get_default(image_object.getPixelSizeY(), 0), self.default_physical_unit),
                                   (get_default(image_object.getPixelSizeZ(), 0), self.default_physical_unit)]
-        objective_settings = image_object.getObjectiveSettings()
-        if objective_settings:
-            self.source_mag = objective_settings.getObjective().getNominalMagnification()
-        else:
-            self.source_mag = 0
         self.channels = []
         for channeli, channel0 in enumerate(image_object.getChannels()):
             channel = {'label': get_default(channel0.getName(), str(channeli)),